moved state of stats to card details
