+++ src/components/CardDetails.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { Card, Message, Statistic, Grid } from 'semantic-ui-react'
+import { BACKEND } from "../App";
 
 //TODO: allow form that lets mentors edit their time preferences
 export default class CardDetails extends React.Component {
@@ -7,6 +8,39 @@ export default class CardDetails extends React.Component {
         super(props);
         this.renderAboutMe = this.renderAboutMe.bind(this);
         this.getLanguages = this.getLanguages.bind(this);
+        this.state = {
+            calls_requested: null,
+            calls_completed: null,
+        }
+    }
+
+    componentDidMount() {
+        if (this.props.isMentor) {
+            if (this.props.mentorIdForStats) {
+                let payload = {
+                    id: this.state.mentorIdForStats
+                }
+                fetch(`${BACKEND}/getRequestRecordsMentor`, {
+                    method: 'post',
+                    headers: {'Content-Type':'application/json'},
+                    body: JSON.stringify(payload)
+                    }).then(async res => {
+                        let resolvedRes = await res;
+                        let status = resolvedRes.status;
+                        resolvedRes = await resolvedRes.json()
+                        if (status !== 200) {
+                        console.log("Could not fetch mentor statistics.");
+                        }
+                        else {
+                            console.log("statistics fetched are", resolvedRes)
+                        this.setState({
+                            calls_completed: resolvedRes.calls_completed,
+                            calls_requested: resolvedRes.calls_requested,
+                        })
+                        }
+                    })
+            }
+        }
     }
 
     renderAboutMe() {
@@ -30,7 +64,6 @@ export default class CardDetails extends React.Component {
 
     render() {
         let {name, school, memberSince} = this.props;
-        console.log("props in card", this.props)
         return (
             <div>
                 <Card>
